Regular Commit
Stuff is actually being saved to the database now. Still need to implement the file-uploading and comment box. After that I need to worry about saving the variables for users even when they log off. The variables shall be nulled every time a user clicks the check out button too.